Atualizado todas as Classes e deixado o código mais limpo. Adicionado algumas funções.

+++ src/ProjetoCRM/Estoque.java
@@ -2,6 +2,6 @@ package ProjetoCRM;
 
 public class Estoque {
     float valorEstoque;
-    float volumeatualEstoque;
+    int volumeatualEstoque;
     int qtdMercadoria;
 }
